Task/cleanup (#5)
* Removed vtkm backend pipeline. Building lodepng as library object and inlcuding it in alpine lib

* Removing last mention of eavl from the code
