+++ src/alpine/alpine_config.h.in
@@ -59,10 +59,6 @@
 #cmakedefine ALPINE_USE_OPENMP        "@OPENMP_FOUND@"
 
 // defs for pipeline support based on which 3rd-party libs we have 
-#cmakedefine ALPINE_EAVL_ENABLED      "@EAVL_FOUND@"
-#cmakedefine ALPINE_EAVL_USE_OPENMP   "@OPENMP_FOUND@"
-#cmakedefine ALPINE_EAVL_USE_CUDA     "@CUDA_FOUND@"
-
 
 #cmakedefine ALPINE_VTKM_ENABLED      "@VTKM_FOUND@"
 #cmakedefine ALPINE_VTKM_USE_CUDA     "@CUDA_FOUND@"
