+++ src/tests/alpine/t_alpine_render_3d.cpp
@@ -129,155 +129,6 @@ TEST(alpine_render_3d, test_render_3d_render_default_pipeline)
     EXPECT_TRUE(check_test_image(output_file));
 }
 
-//-----------------------------------------------------------------------------
-TEST(alpine_render_3d, test_render_3d_render_eavl_serial_backend)
-{
-    Node n;
-    alpine::about(n);
-    // only run this test if alpine was built with eavl support
-    if(n["pipelines/eavl/status"].as_string() == "disabled")
-    {
-        ALPINE_INFO("EAVL support disabled, skipping 3D EAVL-serial test");
-        return;
-    }
-    
-    ALPINE_INFO("Testing 3D Rendering with EAVL Pipeline using Serial Backend");
-
-    
-    //
-    // Create an example mesh.
-    //
-    Node data, verify_info;
-    conduit::blueprint::mesh::examples::braid("hexs",
-                                              EXAMPLE_MESH_SIDE_DIM,
-                                              EXAMPLE_MESH_SIDE_DIM,
-                                              EXAMPLE_MESH_SIDE_DIM,
-                                              data);
-
-    
-    EXPECT_TRUE(conduit::blueprint::mesh::verify(data,verify_info));
-    verify_info.print();
-
-
-    string output_path = prepare_output_dir();
-    string output_file = conduit::utils::join_file_path(output_path, "tout_render_3d_eval_serial_backend");
-
-    // remove old images before rendering
-    remove_test_image(output_file);
-
-    //
-    // Create the actions.
-    //
-
-    Node actions;
-    
-    Node &plot = actions.append();
-    plot["action"]     = "add_plot";
-    plot["field_name"] = "braid";
-
-    Node &opts = plot["render_options"];
-    opts["width"]  = 500;
-    opts["height"] = 500;
-    opts["file_name"] = output_file;
-    
-    actions.append()["action"] = "draw_plots";
-    
-    //
-    // Run Alpine
-    //
-    
-    Node open_opts;
-    open_opts["pipeline/type"] = "eavl";
-    open_opts["pipeline/backend"] = "serial";
-    
-    Alpine sman;
-    sman.Open(open_opts);
-    sman.Publish(data);
-    sman.Execute(actions);
-    sman.Close();
-
-    // check that we created an image
-    EXPECT_TRUE(check_test_image(output_file));
-}
-
-
-//-----------------------------------------------------------------------------
-TEST(alpine_render_3d, test_render_3d_render_eavl_cuda_backend)
-{
-    Node n;
-    alpine::about(n);
-    // only run this test if alpine was built with eavl support
-    if(n["pipelines/eavl/status"].as_string() == "disabled")
-    {
-        ALPINE_INFO("EAVL support disabled, skipping 3D EAVL-cuda test");
-        return;
-    }
-    
-    if(n["pipelines/eavl/backends/cuda"].as_string() != "enabled")
-    {
-        ALPINE_INFO("EAVL CUDA support disabled, skipping 3D EAVL-cuda test");
-        return;
-    }
-
-    ALPINE_INFO("Testing 3D Rendering with EAVL Pipeline using CUDA Backend");
-
-    
-    //
-    // Create an example mesh.
-    //
-    Node data, verify_info;
-    conduit::blueprint::mesh::examples::braid("hexs",
-                                              EXAMPLE_MESH_SIDE_DIM,
-                                              EXAMPLE_MESH_SIDE_DIM,
-                                              EXAMPLE_MESH_SIDE_DIM,
-                                              data);
-
-    
-    EXPECT_TRUE(conduit::blueprint::mesh::verify(data,verify_info));
-    verify_info.print();
-
-
-    string output_path = prepare_output_dir();
-    string output_file = conduit::utils::join_file_path(output_path, "tout_render_3d_eval_cuda_backend");
-
-    // remove old images before rendering
-    remove_test_image(output_file);
-
-    //
-    // Create the actions.
-    //
-
-    Node actions;
-    
-    Node &plot = actions.append();
-    plot["action"]     = "add_plot";
-    plot["field_name"] = "braid";
-
-    Node &opts = plot["render_options"];
-    opts["width"]  = 500;
-    opts["height"] = 500;
-    opts["file_name"] = output_file;
-    
-    actions.append()["action"] = "draw_plots";
-    
-    //
-    // Run Alpine
-    //
-    
-    Node open_opts;
-    open_opts["pipeline/type"] = "eavl";
-    open_opts["pipeline/backend"] = "cuda";
-    
-    Alpine sman;
-    sman.Open(open_opts);
-    sman.Publish(data);
-    sman.Execute(actions);
-    sman.Close();
-
-    // check that we created an image
-    EXPECT_TRUE(check_test_image(output_file));
-}
-
 
 //-----------------------------------------------------------------------------
 TEST(alpine_render_3d, test_render_3d_render_vtkm_serial_backend)
