+++ src/thirdparty_builtin/CMakeLists.txt
@@ -42,16 +42,3 @@
 # 
 ###############################################################################
 
-
-################################
-# Add lodepng 
-################################
-add_subdirectory(lodepng)
-
-
-
-
-
-
-
-
